% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndPrepareDataWB.R
\name{addBiomassDeltas}
\alias{addBiomassDeltas}
\title{Get biomass deltas}
\usage{
addBiomassDeltas(d)
}
\arguments{
\item{d}{dataframe created with getCoreData()}
}
\value{
a data frame including meanBiomassAllSppStdDelta (changes in biomass with bioomass by sample standardized by all species) and meanBiomassStdDelta (changes in biomass with bioomass by sample standardized by each species)
}
\description{
Get biomass deltas
}
