% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthModelFunctions.R
\name{addDensityData}
\alias{addDensityData}
\title{Add density data}
\usage{
addDensityData(ddddG, ddD, ddddD, meanOrIterIn, sampleToUse)
}
\arguments{
\item{drainage}{Which drainage, "west" or "stanley"}
}
\value{
a data frame
}
\description{
Add density data
}
