% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndPrepareDataWB.R
\name{addNPasses}
\alias{addNPasses}
\title{Get pass data from raw data table}
\usage{
addNPasses(cd, dr)
}
\arguments{
\item{drainage}{Which drainage, "west" or "stanley"}
}
\value{
a data frame
}
\description{
Get pass data from raw data table
}
