% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelCase.R
\name{camelCase}
\alias{camelCase}
\title{Convert string of separated words to camel-case}
\usage{
camelCase(x, sep = "_")
}
\arguments{
\item{x}{Character vector to be converted to camel-case}

\item{sep}{Character separating words in the original string}
}
\value{
Camel-case string(s)
}
\description{
Convert string of separated words to camel-case
}
