% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndPrepareDataWB.R
\name{getCoreData}
\alias{getCoreData}
\title{Extract data from the PIT tag database}
\usage{
getCoreData(drainage = "west")
}
\arguments{
\item{drainage}{Which drainage, "west" or "stanley"}
}
\value{
a data frame
}
\description{
Extract data from the PIT tag database
}
