% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeOutputFunctions.R
\name{getPrediction}
\alias{getPrediction}
\title{Generate predictions from a set of iterations of a jags model run}
\usage{
getPrediction(d, limits = 2, nPoints = 5, itersForPred, varsToEstimate)
}
\arguments{
\item{d}{a model run dataFrame}

\item{limits}{symetrical lower and upper limit for the predictions}

\item{nPoints}{the number of points between the limits. Use an odd # to ensure a value of 0}

\item{itersForPred, }{the iterations from the model for prediction}

\item{varsToEstimate}{which variables to use for estimates (len,count,phi,temp,flow)}
}
\value{
a data frame
}
\description{
Generate predictions from a set of iterations of a jags model run
}
