% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeOutputFunctions.R
\name{getPredictionSigma}
\alias{getPredictionSigma}
\title{Generate predictions of sigma the sd in growth from a set of iterations of a jags model run}
\usage{
getPredictionSigma(d, limits = 2, nPoints = 5, itersForPred)
}
\arguments{
\item{d}{a model run dataFrame}

\item{limits}{symetrical lower and upper limit for the predictions}

\item{nPoints}{the number of points between the limits. Use an odd # to ensure a value of 0}

\item{itersForPred, }{the iterations from the model for prediction}
}
\value{
a data frame
}
\description{
Generate predictions of sigma the sd in growth from a set of iterations of a jags model run
}
