% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndPrepareDataWB.R
\name{getPropSampled}
\alias{getPropSampled}
\title{Get propoortion of sections sampled by river,sample}
\usage{
getPropSampled(nSeasons, nRivers, nYears)
}
\value{
a list of propSampleDATA, and zeroSectionsDATA
}
\description{
Get propoortion of sections sampled by river,sample
}
