% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeOutputFunctions.R
\name{getRMSE}
\alias{getRMSE}
\title{Plot observed/predicted and return RMSE}
\usage{
getRMSE()
}
\value{
Nothing, print text and plot graph
}
\description{
Plot observed/predicted and return RMSE
}
