% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndPrepareDataWB.R
\name{getSites}
\alias{getSites}
\title{Get data from sites table}
\usage{
getSites(drainageIn = "west")
}
\arguments{
\item{drainage}{Which drainage, "west" or "stanley"}
}
\value{
a data frame
}
\description{
Get data from sites table
}
