% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndPrepareDataWB.R
\name{getYOYCutoffs}
\alias{getYOYCutoffs}
\title{Get cutoFFYoy data}
\usage{
getYOYCutoffs(d, dr)
}
\arguments{
\item{d}{dataframe created with getCoreData()}

\item{dr}{Which drainage, "west" or "stanley"}
}
\value{
a data frame
}
\description{
Get cutoFFYoy data
}
