% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndPrepareDataWB.R
\name{mergeSites}
\alias{mergeSites}
\title{Merge sites table}
\usage{
mergeSites(d, drainageIn)
}
\arguments{
\item{d}{dataframe created with getCoreData()}

\item{drainage}{Which drainage, "west" or "stanley"}
}
\value{
a data frame
}
\description{
Merge sites table
}
