% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndPrepareDataWB.R
\name{minimalData}
\alias{minimalData}
\title{Show minimal data}
\usage{
minimalData(d)
}
\arguments{
\item{d}{dataframe created with getCoreData()}
}
\value{
a data frame with key selected columns
}
\description{
Show minimal data
}
