% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeOutputFunctions.R
\name{plotPred}
\alias{plotPred}
\title{Plot predictions from a set of iterations of a jags model run}
\usage{
plotPred(p, varsToPlot, isYOYGG, speciesGG)
}
\arguments{
\item{p}{a model run set of predictions from getPredictions()}

\item{varsToPlot}{variables for ploting. Eith singles or pairs of (flow,temp,count)}

\item{isYOYGG, }{isYOY, 0 or 1}

\item{speciesGG, }{bkt, bnt or ats}
}
\value{
a ggplot object
}
\description{
Plot predictions from a set of iterations of a jags model run
}
