% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectionModelFunctions.R
\name{runDetectionModel}
\alias{runDetectionModel}
\title{Run the detection model}
\usage{
runDetectionModel(d, parallel = FALSE)
}
\arguments{
\item{drainage}{Which drainage, "west" or "stanley"}
}
\value{
a data frame
}
\description{
Run the detection model
}
