% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthModelFunctions.R
\name{addDensityData}
\alias{addDensityData}
\title{Add density and biomass data}
\usage{
addDensityData(ddddGIn, ddDIn, ddddDIn, meanOrIterIn, sampleToUse)
}
\arguments{
\item{ddddGIn}{growth data}

\item{ddDIn}{results from the detection model}

\item{meanOrIterIn}{whether the growth model gets mean P from the detection model, or results from an iteration}

\item{sampleToUse}{if detection results are from an iteration, which iteration}

\item{ddddGIn}{detection data}
}
\value{
a data frame with standardized counts, countPStd is standardized by species counts, countPAllSppStd is standardized by the sum of counts of all species in the analysis
}
\description{
Add density and biomass data
}
