% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeOutputFunctions.R
\name{getPrediction}
\alias{getPrediction}
\title{Generate predictions from an iteration of a jags model run}
\usage{
getPrediction(d, limits = 2, nPoints, iterForPred)
}
\arguments{
\item{d}{a model run dataFrame}

\item{limits}{symetrical lower and upper limit for the predictions}

\item{nPoints}{the number of points between the limits. Use an odd # to ensure a value of 0}

\item{iterForPred, }{the iteration from the model for prediction}
}
\value{
a data frame
}
\description{
Generate predictions from an iteration of a jags model run
}
