% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareDataForJags.R
\name{prepareDataForJags}
\alias{prepareDataForJags}
\title{Get data ready to run the detection or growth model}
\usage{
prepareDataForJags(d, modelType)
}
\arguments{
\item{d}{a data frame created using getCoreData()}

\item{modelType}{a character variable specifying the model type, ("detection', "growth"). This specifies the data that go into 'data'}
}
\value{
A list for importing into the jags run. Run with run"modelType"Model()
}
\description{
Get data ready to run the detection or growth model
}
