% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareDataForJags.R
\name{prepareDataForJags}
\alias{prepareDataForJags}
\title{Get data ready to run the growth model}
\usage{
prepareDataForJags(d)
}
\arguments{
\item{d}{a data frame created using getCoreData()}
}
\value{
A list for importing into the jags run. Run with runGrowthModel()
}
\description{
Get data ready to run the growth model
}
