% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthModelFunctions.R
\name{runGrowthModel}
\alias{runGrowthModel}
\title{Run the growth model using jags}
\usage{
runGrowthModel(d, parallel = FALSE)
}
\arguments{
\item{d, }{the input dataframe}

\item{paralel, }{boolean for running chains in parallel}
}
\value{
a data frame
}
\description{
Run the growth model using jags
}
