% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthModelFunctions.R
\name{runGrowthModel}
\alias{runGrowthModel}
\title{Run the growth model}
\usage{
runGrowthModel(d, parallel = FALSE)
}
\arguments{
\item{drainage}{Which drainage, "west" or "stanley"}
}
\value{
a data frame
}
\description{
Run the growth model
}
